% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessary_code.R
\name{calculate_and_write_pairwise_dmi}
\alias{calculate_and_write_pairwise_dmi}
\title{Calculate all pairwise RDI and spearman rank correlation}
\usage{
calculate_and_write_pairwise_dmi(genes_data, delays = c(1, 2, 5, 10, 15, 20,
  25), supergraph = NULL, cores = 1, verbose = T)
}
\arguments{
\item{genes_data}{pseudo-time/time series for the gene expression data}

\item{delays}{a vector of delays to scanning the data}

\item{supergraph}{a dataframe storing the list of possible edges}

\item{cores}{number of cores used for performing the calculation}

\item{verbose}{a logic argument to determine whether or not the details of calculation should be reported}
}
\value{
a dataframe storing the RDI results. The first two columns correspond to the ensemble id while the last column corresponds to the spearman
rank correlation. The rest columns corresponds to the RDI values under different delays
}
\description{
This function estimates the RDI and the spearman correlation of expression (expr1, expr2) for any two genes, id1, di2 in the pseudo-time-series data
}

