% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessary_code.R
\name{calculate_and_write_pairwise_dmi_conditioned}
\alias{calculate_and_write_pairwise_dmi_conditioned}
\title{Calculate all pairwise conditional RDI values}
\usage{
calculate_and_write_pairwise_dmi_conditioned(genes_data, rdi_res, k = 1,
  supergraph = NULL, cores = 1, verbose = T)
}
\arguments{
\item{genes_data}{expression matrix for all the genes included in the RDI calculation}

\item{rdi_res}{calculated RDI results, including ids under test, RDI values under different delays and spearman rank corelation}

\item{k}{number of top k-incoming nodes for calculating RDI values. Default is 1. Higher k requres to have more samples or longer time / "pseudotime" series}

\item{supergraph}{a data frame stores the list of feasible edges for RDI analysis. Possible edges outside of this graph will be ignored}

\item{cores}{number of cores used in the calculation}

\item{verbose}{a logical argument to determine wheter or not to print the detailed running information.}
}
\value{
a dataframe storing conditional RDI results. First two columns are the id names for the genes.
Third column is the conditional RDI value and last column is the time delay corresponding to the largest RDI.
}
\description{
This function estimates the RDI value for every gene-pair combination conditioned on the top K incoming edges.
Delay for the gene-pair under test corresponds to the delay with highest RDI value while delay for the conditioned K
incoming edges are based on their corresponding delay value.
}

