% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessary_code.R
\name{calculate_conditioned_rdi}
\alias{calculate_conditioned_rdi}
\title{Calculate conditionally RDI value}
\usage{
calculate_conditioned_rdi(id1, id2, genes_data, d, top_k_genes_delay,
  N_operations)
}
\arguments{
\item{id1}{ensemble id for gene 1}

\item{id2}{ensemble id for gene 2}

\item{genes_data}{pseudo-time/time series for the gene expression data}

\item{d}{delay for current gene interaction under test corresponding to the largest RDI value}

\item{top_k_genes_delay}{a vector of top k incoming node's delay corresponding to the maximal RDI value}

\item{N_operations}{number of total edges for calculating RDI values (not used for now)}
}
\value{
a dataframe storing conditional RDI results. First two columns are the id names for the genes.
Third column is the conditional RDI value.
}
\description{
This function estimates the RDI value for gene-pair combination (id1, id2) in the pseudotime/time series data, conditioned on the top K incoming edges.
Delay for the gene-pair under test corresponds to the delay with highest RDI value while delay for the conditioned k
incoming edges are based on their corresponding delay value.
}

