% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessary_code.R
\name{calculate_rdi_corr}
\alias{calculate_rdi_corr}
\title{RDI and the spearman correlation of expression}
\usage{
calculate_rdi_corr(id1, id2, expr1, expr2, delays, N_operations)
}
\arguments{
\item{id1}{ensemble id for gene 1}

\item{id2}{ensemble id for gene 2}

\item{expr1}{number for nearest neighbors used in entropy calculation}

\item{expr2}{number for nearest neighbors used in entropy calculation}

\item{delays}{number for nearest neighbors used in entropy calculation}

\item{N_operations}{number for nearest neighbors used in entropy calculation}
}
\value{
a vector of entropy values
}
\description{
This function estimates the RDI and the spearman correlation of expression (expr1, expr2) for two genes,
id1, di2 along the pseudo-time-series
}

