% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{cmi}
\alias{cmi}
\title{This function estimates the CONDITIONAL mutual information of X and Y given Z}
\usage{
cmi(x, y, z, k = 5)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{condition random variable for variables (x, y) from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}
}
\value{
a numeric value for the condition mutual information estimator between two variables (x, y) conditioned on variable z
}
\description{
This function estimates the CONDITIONAL mutual information of X and Y given Z
}

