% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{di}
\alias{di}
\title{This function calculates Deficit Index (DI) and returns results 
in form of an original data frame with additional DI column.}
\usage{
di(dat, cols, age = NULL, visible = FALSE)
}
\arguments{
\item{dat}{A data frame.}

\item{cols}{A list of column names.}

\item{age}{A name of column which represents age of a patient.
Default: \code{NULL}.}

\item{visible}{A flag to show DI plot (mean DI in a population by age)
Default: \code{FALSE}}
}
\value{
An original data frame with additional DI column.
}
\description{
This function calculates Deficit Index (DI) and returns results 
in form of an original data frame with additional DI column.
}
\examples{
library(di)
ddi <- di(dd, di_columns, age="Age_orig", visible = T)
}
