\name{di-package}
\alias{di-package}
\alias{di}
\docType{package}
\title{
\packageTitle{di}
}
\description{
\packageDescription{di}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{di}
\packageIndices{di}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{di}

Maintainer: \packageMaintainer{di}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
