% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{di_single_run}
\alias{di_single_run}
\title{This function simulates the DIRECTED mutual information from X to Y when you have a SINGLE run of the process}
\usage{
di_single_run(x, y, n = 10)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}
}
\value{
a matrix for the condition mutual information estimators between all pairwise variables (x, y) in the data matrix x, y
}
\description{
This function simulates the DIRECTED mutual information from X to Y when you have a SINGLE run of the process
}

