% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{entropy}
\alias{entropy}
\title{This function estimates the entropy of a continuous random variable}
\usage{
entropy(x, k = 5)
}
\arguments{
\item{x}{data matrix used for calculating the entropy}

\item{k}{number for nearest neighbors used in entropy calculation}
}
\value{
a vector of entropy values
}
\description{
This function estimates the entropy of a continuous random variable
}

