% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{mi}
\alias{mi}
\title{This function estimates the mutual information of two random variables based on their observed values}
\usage{
mi(x, y, k = 5)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}
}
\value{
a numeric value for the mutual information estimator between two variables (x, y)
}
\description{
This function estimates the mutual information of two random variables based on their observed values
}

