% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{rdi_single_run_conditioned}
\alias{rdi_single_run_conditioned}
\title{This function simulates the CONDITIONED DIRECTED mutual information from X to Y CONDITIONED ON Z when you have a SINGLE run of the processes}
\usage{
rdi_single_run_conditioned(x, y, z, z_delays, d = 1)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{z is a dataframe or matrix consisting of the data for different variables}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}

\item{z_delay}{z_delay is also a dataframe or matrix consisting of the delays to be applied to different variables}
}
\value{
a matrix for the condition mutual information estimators between all pairwise variables (x, y) in the data matrix x, y
}
\description{
This function simulates the CONDITIONED DIRECTED mutual information from X to Y CONDITIONED ON Z when you have a SINGLE run of the processes
}

