% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessary_code.R
\name{smooth_genes}
\alias{smooth_genes}
\title{This function smoothes the data using moving average}
\usage{
smooth_genes(exprs, window_size = 40)
}
\arguments{
\item{exprs}{a dataframe or matrix to be smoothed, row corresponds variables while column correspond to sample}

\item{window_size}{size of sliding window}
}
\value{
moving average smoothed dataset
}
\description{
This function smoothes the data using moving average
}

