% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{vd}
\alias{vd}
\title{This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm}
\usage{
vd(d)
}
\arguments{
\item{d}{number of deminsion}
}
\value{
a numeric value for the d-dimensional unit ball for Euclidean norm
}
\description{
This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm
}

